# openapi-spec.yaml
openapi: 3.0.3
info:
  title: MyLab Platform API
  description: Complete API specification for the MyLab Platform backend
  version: 1.1.0
  contact:
    name: MyLab Platform Team
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.mylab-platform.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /samples:
    get:
      summary: List samples
      description: Get workspace samples
      tags: [Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, discarded]
      responses:
        '200':
          description: Samples list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Sample'

    post:
      summary: Create sample
      description: Create new sample
      tags: [Samples]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSampleRequest'
      responses:
        '201':
          description: Sample created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Sample'
  /samples/{id}:
    get:
      summary: Get sample
      description: Get sample details
      tags: [Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sample details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Sample'

    put:
      summary: Update sample
      description: Update sample
      tags: [Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSampleRequest'
      responses:
        '200':
          description: Sample updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Sample'

    delete:
      summary: Delete sample
      description: Delete sample (optionally cascade delete all derived samples with admin permission)
      tags: [Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cascade
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: If true, deletes all derived samples (requires admin role)
      responses:
        '200':
          description: Sample deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================================================
  # DERIVED SAMPLES
  # ============================================================================
  /derived-samples:
    get:
      summary: List derived samples
      description: Get derived samples
      tags: [Derived Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: parentSampleId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Derived samples list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DerivedSample'

    post:
      summary: Create derived sample
      description: Create new derived sample
      tags: [Derived Samples]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDerivedSampleRequest'
      responses:
        '201':
          description: Derived sample created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DerivedSample'
  /derived-samples/{id}:
    get:
      summary: Get derived sample
      description: Get derived sample details
      tags: [Derived Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Derived sample details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DerivedSample'

    put:
      summary: Update derived sample
      description: Update derived sample
      tags: [Derived Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDerivedSampleRequest'
      responses:
        '200':
          description: Derived sample updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DerivedSample'

    delete:
      summary: Delete derived sample
      description: Delete derived sample
      tags: [Derived Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Derived sample deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /derived-samples/{id}/share:
    post:
      summary: Share derived sample
      description: Share derived sample with other users
      tags: [Derived Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userIds
              properties:
                userIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                permissions:
                  type: string
                  enum: [read, write]
                  default: read
      responses:
        '200':
          description: Sample shared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================================================
  # ANALYSES API
  # ============================================================================
...
  /batches/{id}/samples:
    get:
      summary: Get batch samples
      description: Get all samples associated with a batch
      tags: [Batches]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Batch samples
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sample'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Batch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Add sample to batch
      description: Add a sample to an existing batch
      tags: [Batches]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sampleId
              properties:
                sampleId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Sample added to batch successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Batch or sample not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        details:
          type: object
          description: Additional error details
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            total:
              type: integer
              example: 100
            totalPages:
              type: integer
              example: 5
...
    Sample:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        projectName:
          type: string
        trialId:
          type: string
          format: uuid
        stageId:
          type: string
          format: uuid
        sampleId:
          type: string
          description: User-defined sample identifier
        description:
          type: string
        type:
          type: string
          description: Sample type/category
        metadata:
          type: object
          description: Additional sample metadata
        status:
          type: string
          enum: [active, archived, discarded]
          default: active
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
    CreateSampleRequest:
      type: object
      required:
        - projectId
        - sampleId
        - description
      properties:
        projectId:
          type: string
          format: uuid
        stageId:
          type: string
          format: uuid
        trialId:
          type: string
          format: uuid
        sampleId:
          type: string
        description:
          type: string
        type:
          type: string
        metadata:
          type: object
    UpdateSampleRequest:
      type: object
      properties:
        sampleId:
          type: string
        description:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [active, archived, discarded]
        metadata:
          type: object
        stageId:
          type: string
          format: uuid
        trialId:
          type: string
          format: uuid

    # ============================================================================
    # DERIVED SAMPLE SCHEMAS
    # ============================================================================
    DerivedSample:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parentSampleId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        derivationMethod:
          type: string
        metadata:
          type: object
        status:
          type: string
          enum: [active, archived, discarded]
          default: active
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
    CreateDerivedSampleRequest:
      type: object
      required:
        - parentSampleId
        - name
        - derivationMethod
      properties:
        parentSampleId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        derivationMethod:
          type: string
        metadata:
          type: object
    UpdateDerivedSampleRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        derivationMethod:
          type: string
        metadata:
          type: object
        status:
          type: string
          enum: [active, archived, discarded]

    # ============================================================================
    # ANALYSIS SCHEMAS
    # ============================================================================
...
    Batch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        batchNumber:
          type: string
          description: Unique batch identifier
        name:
          type: string
        description:
          type: string
        batchType:
          type: string
          enum: [production, qc, research, validation]
        status:
          type: string
          enum: [draft, active, completed, archived]
          default: draft
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        targetCompletionDate:
          type: string
          format: date
        actualCompletionDate:
          type: string
          format: date
        metadata:
          type: object
          description: Additional batch metadata
        sampleCount:
          type: integer
          description: Number of samples in the batch
        analysisCount:
          type: integer
          description: Number of analyses in the batch
        assignedTo:
          type: string
          format: uuid
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...